+++ packages/wrapped-keys-lit-actions/src/lib/raw-action-functions/solana/signMessageWithEncryptedSolanaKey.js
@@ -1,0 +1,51 @@
+const {
+  getDecryptedKeyToSingleNode,
+} = require('../../internal/common/getDecryptedKeyToSingleNode');
+const { signMessageSolanaKey } = require('../../internal/solana/signMessage');
+const { removeSaltFromDecryptedKey } = require('../../utils');
+
+/* global Lit */
+
+/**
+ *
+ * Bundles solana/web3.js package as it's required to sign a message with the Solana wallet which is also decrypted inside the Lit Action.
+ *
+ * @jsParam pkpAddress - The Eth address of the PKP which is associated with the Wrapped Key
+ * @jsParam ciphertext - For the encrypted Wrapped Key
+ * @jsParam dataToEncryptHash - For the encrypted Wrapped Key
+ * @jsParam messageToSign - The unsigned message to be signed by the Wrapped Key
+ * @jsParam accessControlConditions - The access control condition that allows only the pkpAddress to decrypt the Wrapped Key
+ *
+ * @returns { Promise<string> } - Returns a message signed by the Solana Wrapped key. Or returns errors if any.
+ */
+
+export async function signMessageWithEncryptedSolanaKey({
+  accessControlConditions,
+  ciphertext,
+  dataToEncryptHash,
+  messageToSign,
+}) {
+  try {
+    const decryptedPrivateKey = await getDecryptedKeyToSingleNode({
+      accessControlConditions,
+      ciphertext,
+      dataToEncryptHash,
+    });
+
+    if (!decryptedPrivateKey) {
+      // Silently exit on nodes which didn't run the `decryptToSingleNode` code
+      return;
+    }
+
+    const privateKey = removeSaltFromDecryptedKey(decryptedPrivateKey);
+
+    const signature = await signMessageSolanaKey({
+      messageToSign,
+      privateKey,
+    });
+
+    Lit.Actions.setResponse({ response: signature });
+  } catch (err) {
+    Lit.Actions.setResponse({ response: `Error: ${err.message}` });
+  }
+}
